% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list.R
\name{add_items_if_not_present}
\alias{add_items_if_not_present}
\title{Add named arguments to the current list only if they do not already exist}
\usage{
add_items_if_not_present(current, update)
}
\arguments{
\item{current}{current list}

\item{update}{named list of items to update}
}
\value{
new list with elements of update added to current if there isn't
already an element of that name
}
\description{
Add named arguments to the current list only if they do not already exist
}
