% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{evaluate_call}
\alias{evaluate_call}
\title{Evaluate a call with options for changing the arguments}
\usage{
evaluate_call(
  x,
  defaults = list(),
  update = list(),
  delete = character(0),
  envir = parent.frame()
)
}
\arguments{
\item{x}{call object}

\item{defaults}{list of arguments which would act as the default. These will
be superseded by any arguments of the same name that already exist
in the call object \code{x}}

\item{update}{named list of values which override arguments in the call.
Note: "NULL" is \emph{not} used to remove an argument, but will actually
set that argument value to NULL.  If you want to remove an argument, use
\code{delete}.}

\item{delete}{character vector of named arguments to delete.}

\item{envir}{evaluation environment}
}
\description{
This is an augmented version of \code{base::eval()} which has options for
overriding arguments before evaluation
}
