% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{is_fully_named_list}
\alias{is_fully_named_list}
\title{Check a list is either empty or fully named}
\usage{
is_fully_named_list(ll)
}
\arguments{
\item{ll}{list}
}
\value{
logical
}
\description{
Check a list is either empty or fully named
}
