% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{modify_call}
\alias{modify_call}
\title{Change a call by updating its arguments}
\usage{
modify_call(x, defaults = list(), update = list(), delete = character(0))
}
\arguments{
\item{x}{call object}

\item{defaults}{list of arguments which would act as the default. These will
be superseded by any arguments of the same name that already exist
in the call object \code{x}}

\item{update}{named list of values which override arguments in the call.
Note: "NULL" is \emph{not} used to remove an argument, but will actually
set that argument value to NULL.  If you want to remove an argument, use
\code{delete}.}

\item{delete}{character vector of named arguments to delete.}
}
\description{
Change a call by updating its arguments
}
\examples{
cc <- capture_call(farnarkle(6, x = 2, z = 9))
cc <- modify_call(cc, defaults = list(x = 1, y = 7), update = list(z = 0))
cc


}
