% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list.R
\name{modify_list}
\alias{modify_list}
\title{Add named arguments to the current list}
\usage{
modify_list(current, update)
}
\arguments{
\item{current}{current list}

\item{update}{named list of items to update}
}
\value{
new list with all elements of update added to current
}
\description{
Add named arguments to the current list
}
